+++ WorkloadManagementSystem/Utilities/JobParameters.py
@@ -57,7 +57,7 @@ def getMemoryFromProc():
   meminfo = dict((i.split()[0].rstrip(':'), int(i.split()[1])) for i in open('/proc/meminfo').readlines())
   maxRAM = meminfo['MemTotal']
   if maxRAM:
-    return maxRAM / 1024
+    return int(maxRAM / 1024)
 
 
 def getNumberOfProcessors(siteName=None, gridCE=None, queue=None):
